pyproject.toml
[tool.poetry]
name = "dash-pts"
version = "0.1.0"
description = ""
authors = ["dunossauro <mendesxeduardo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
dash = "^1.19.0"
Flask-SQLAlchemy = "^2.5.1"
pendulum = "^2.1.2"

[tool.poetry.dev-dependencies]
behave = "^1.2.6"
selenium = "^3.141.0"
black = "^20.8b1"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
